Set title to filename
Most browsers, when you press Ctrl+S will attempt to save the HTML file with
the title as the base-name, followed by .html. This little commit makes sure
to set the title to whatever the current filename of the HTML page is, such
that Ctrl+S will overwrite the original file.